requirements/prod.txt
panini==0.2.3
PyYAML==5.1.2
python-dotenv==0.17.0
